// frontend/src/api/generated/models/DeviceSettingsResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DeviceSettingsResponse = {
    host: string;
    port: string;
    username: string;
    /**
     * Always redacted as ********
     */
    password: string;
    fromEmail: string;
    recipientEmail: string;
};